+++ ISSUE.md
@@ -1,19 +1,5 @@
 This ISSUE file is dedicated to all the current issues in this development branch.
 
 # Commands
-> All commands listed under this; are refactored with new code, but have some issues. And ISSUE is mentioned in \<command>.py file only, with format 
 
-```py
-    """
-    ISSUE:
-        "issue mentioned here"
-
-    """
-```
-
-- commands.general.math
-- commands.general.movie
-- commands.general.search
-- commands.game.poll
-
-*All API commands, need to be refactored by someone who has api keys.*
+- command.database
