+++ command/database/loader/__init__.py
@@ -1,0 +1,4 @@
+# init, importing everything from loader.py
+
+from .loader import *
+from .database import *
