+++ command/database/rank.py
@@ -1,14 +1,17 @@
+import loader
 from discord.ext import commands
 
-class Ping(commands.Cog):
+
+class Rank(commands.Cog):
     def __init__(self, client):
         self.client = client
 
     @commands.command()
-    async def ping(self, ctx):
-
-        await ctx.send(f"🏓 Pong: `{round(self.client.latency * 1000)}ms`")
+    async def rank(self, ctx):
+        db = loader.db_loaded()
+        await db.rank_query(ctx.message)
         await ctx.message.add_reaction("\U0001f44d")
 
+
 def setup(client):
-    client.add_cog(Ping(client))
+    client.add_cog(Rank(client))
