+++ command/general/search.py
@@ -1,3 +1,4 @@
+import json
 import discord
 from discord.ext import commands
 from duckduckgo_search import ddg
@@ -15,23 +16,16 @@ class Search(commands.Cog):
 
 
     @commands.command()
-    async def search(self, ctx):
-        hold = ctx.message.content.find(" ")
-        results = str(
-            ddg(
-                ctx.message.content[(hold + 1) : len(ctx.message.content)],
-                region="wt-wt",
-                safesearch="Off",
-                time="y",
-                max_results=1,
-            )
-        )
-        index = results.find("'body'")
+    async def search(self, ctx, query):
+
+        results = ddg(query, region = 'wt-wt', safesearch = 'Off', time = 'y', max_results = 1)[0]
+        data = json.loads(str(results).replace("'", '"'))
+        
         embed = discord.Embed(
-            title="Search results for : "
-            + (ctx.message.content[(hold + 1) : len(ctx.message.content)]),
-            description=results[index + 9 : (len(results) - 3)],
-            color=discord.Color.blue(),
+            title = data['title'],
+            color = discord.Color.blue(),
+            description = data['body'],
+            url = data['href'],
         )
 
         await ctx.send(embed=embed)
